Add index.html and style.css and helmet image
Finalized the layout of the page and selected a color theme. Still not
ready to go live, yet. I want to finalize my bio before that, at the
very least. I would also like to get the JS working to make the page
dynamic, and add in my contact info, some links, and start blogging.

I would also like to get the graphics done, and at least on the page.
The current helmet "logo" is not the final version, and I do have the
suit done (about me). I still need to work on the gloves  for
(projects/links), boots (blog?), and one other design - maybe the
Bio-Medical Comm Harness for the Contact page. I might even keep the
caption for that one (all of them?).