Add new images, styles, custom fonts
Tweaked the helmet image in top right, adding in the caption and making
it white for visibility. Changed the title from a somewhat blurry image
to actual text. Start getting ready to add custom fonts - title is
Century Gothic, which should be okay, though I probably should make
sure of that. Caption text (and title H2) is a custom font called TGL
that really nicely simulates the technical manual font.

I also adjusted the header styles so that my name/title aligns with the
main body, while the helmet is centered over the nav.

And, I added a .js file that should allow me to dynamically change
what's displayed in the main body section.