Fix responsive spacing, tweak nav bar
I'd noticed that when viewing on a cell phone-sized screen, the images
appropriately switched to the bottom of the screen, but WAY down. Turns
out, I'd set the inner container min-height to 100vh, when I needed to
set this on the outer container. I also made sure that the contact page
image didn't swap places, as unlike the other pages, it is below the
text to begin with.

The nav bar I adjusted so that it is a set of paragraphs in a div,
rather than an ordered list. This gives me more control over the
labels, but also allows me to highlight the current page.