Add JS to load the nav menu to each page
Instead of having the UL on each page, I added some JavaScript to add
the list automatically to each page, so that I don't have to change
each page every time I change something. I also added a bit of code
to the javascript to highlight the current page in the list.