Adjust font sizes and header image
Using clamp(), I made the font sizes for all of the text on the page
variable, depending on the size of the window. I also tweaked the
title, subtitle, body h1 and the nav text so that none of it will wrap.
This shouldn't really be an issue, as the font should shrink to fit the
space. I was finding, though, that the text would break before
shrinking. With the text being justified, this lead to wonky spacing.
So, I forced it not to break.

I also limited the width of the body text, so that it doesn't just
stretch across the screen, making it more readable.

I think the background is gorked now, though, so I'll need to work on
that next (as well as verifying I don't need most of the currently
commented-out media queries so that I can delete them).