Adjust more of the font sizes
Cleaned up more of the font sizes so that everything continues to look
right/good at smaller screen sizes. The main title was tricky, as it
didn't want to scale the way I wanted it to, and there was some
conflict with the helmet image, as it hides under a certain screen
width, which changes the amount of space for the title (and so, it can
be bigger). I was also struggling to get the body text to "fit" at
320px width. It kept overflowing the box, and I'm not sure why. I
reduced the width of the nav sidebar (and the nav fonts), and that
seemed to do the trick.