Simplify the title and adjust body layout
After checking the live site on a couple of devices, I've decided to
remove the helemt image from the header entirely. I really like the
image, so I moved it to the Contact page in place of the comm harness.
In order to keep the title centered, and not stretched across the
entire page, I added a couple of empty divs on either side of the title
and then set their displays to none at narrower widths.

I also adjusted the body text again, moving the images back to the
right side, and chaging the spacing to 'justify-content: space-evenly;'
to add a little whitespace around the text.