Add some custom variables for the colors
I wanted to play with custom variables, and I figured the colors were a
good place to start. Now, all of the background and text colors are
defined at the top of the css file, and can be more easily changed.
Further, the hover and visited link colors are now based on the main
link color, so one change alters all of them.