Shrink the file size of images
For whatever reason, I'd left the images at nearly full size. This made
placing them difficult, and there were larger than they needed to be
(around 3MB each). I shrank them down in Photoshop and changed the src
attributes to point to the smaller images, with links to the full-size
copies. This should allow the page to load a little quicker (not that
this had been a problem, but it might have been).